requirements.txt
Flask
Flask-Cors
PyJWT
Werkzeug
gunicorn
python-dotenv
upstash-redis
requests
twilio
sendgrid
vercel-blob

    